Refactoring transformer to new file, some cleanup
