+++ unit_tests.py
@@ -5,7 +5,7 @@ import tensorflow as tf
 
 import embed
 import utils
-import networks
+import transformers
 
 def static_rnn(core, inputs, initial_state):
   unroll_length = tf.nest.flatten(inputs)[0].shape[0]
@@ -57,9 +57,9 @@ class EmbedTest(unittest.TestCase):
 
     self.assertEqual(embed_game_unflat, embed_game_struct)
 
-class Test_NetworksTransformers(unittest.TestCase):
+class Test_Transformers(unittest.TestCase):
   def test_attention(self):
-    mhab = networks.MultiHeadAttentionBlock(8, 512)
+    mhab = transformers.MultiHeadAttentionBlock(8, 512)
     # Shape grabbed from breakpoint of slippi
     test_query = tf.ones([64, 32, 866]) # This should be the current position
     test_keys = tf.ones([64, 32, 866]) # All in sequence
